+++ Plugins/UnrealMCP/Source/UnrealMCP/Private/EpicUnrealMCPBridge.cpp
@@ -91,6 +91,7 @@
 #include "Commands/EpicUnrealMCPMobileXrCommands.h"
 #include "Commands/EpicUnrealMCPSourceControlCommands.h"
 #include "Commands/EpicUnrealMCPLocalizationCommands.h"
+#include "Commands/EpicUnrealMCPTestingValidationCommands.h"
 #include "Commands/EpicUnrealMCPPhysicsCommands.h"
 #include "Commands/EpicUnrealMCPValidationCommands.h"
 #include "Commands/EpicUnrealMCPInstanceCommands.h"
@@ -364,6 +365,7 @@ void UEpicUnrealMCPBridge::RegisterHandlers()
     RegisterHandler<FEpicUnrealMCPMobileXrCommands>(38); // Mobile / XR (Sub-batch T, route 38, issue #59)
     RegisterHandler<FEpicUnrealMCPSourceControlCommands>(32); // Source Control / Multi-User (Sub-batch U, route 32, issue #60)
     RegisterHandler<FEpicUnrealMCPLocalizationCommands>(33); // Localization (Sub-batch V, route 33, issue #58)
+    RegisterHandler<FEpicUnrealMCPTestingValidationCommands>(39); // Testing / Validation extensions (Sub-batch W, route 39, issue #57)
 }
 
 UEpicUnrealMCPBridge::UEpicUnrealMCPBridge()